requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.3
djangorestframework==3.10.0
idna==2.8
pkg-resources==0.0.0
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
requests-cache==0.5.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
